Update to latest BSP RCW settings
This includes setting the SDHC CD / WP pins to their SDHC
function rather than GPIO, as well as removing the default
fspi divisor and running the spi-nor device at the bootstrap
speed.  This allows the spi-nor flash to bootstrap regardless
of the Bus Speed, or the speed of the flash attached if the
external backup flash is being used.

I also removing the auto boot loading and only including the fspi
pointer copy.  This is a noop if the SD boot is selected, and
allows fspi images to boot correctly.

Signed-off-by: Jon Nettleton <jon@solid-run.com>